requirements/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.0
Django==4.1.4
django-cuser==2017.3.16
django-debug-toolbar==3.2.2
django-extensions==3.2.1
django-rest-knox==4.2.0
django-rest-swagger==2.1.2
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
Pillow==9.3.0
psycopg2-binary==2.9.1
pycparser==2.21
pytz==2022.7
requests==2.28.1
simplejson==3.18.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
